Minor updates, stable and checked :)
